added a very dumb solution for getting mario objects in gfx_pc.c

+++ src/saturn/saturn_actors.h
@@ -26,7 +26,7 @@ extern int saturn_actor_sizeof();
 extern "C" {
 #endif
     void bhv_mario_actor_loop();
-    void override_cc_color(float* r, float* g, float* b, int ccIndex, int marioIndex, int shadeIndex, float intensity, bool additive);
+    void override_cc_color(int* r, int* g, int* b, int ccIndex, int marioIndex, int shadeIndex, float intensity, bool additive);
 #ifdef __cplusplus
 }
 #endif
